show all option in dialog
