+++ src/layouts/i18n/de.ts
@@ -1,0 +1,7 @@
+export default {
+  addDialog: {
+    vehicle: 'Fahrzeug',
+    expense: 'Ausgabe',
+    refuel: 'Betankung'
+  }
+}
