+++ src/layouts/i18n/en.ts
@@ -1,0 +1,7 @@
+export default {
+  addDialog: {
+    vehicle: 'Vehicle',
+    expense: 'Expense',
+    refuel: 'Refuel'
+  }
+}
