+++ src/layouts/i18n/index.ts
@@ -1,0 +1,8 @@
+import en from './en'
+import de from './de'
+
+export default {
+  'en-US': en,
+  en: en,
+  de: de
+}
