+++ src/pages/dashboard/DashboardPage.vue
@@ -89,7 +89,7 @@ import { initSettings } from 'src/scripts/initSettings'
 import { SplashScreen } from '@capacitor/splash-screen'
 import { DashboardData } from './scripts/models'
 import { ThemeSetter } from 'src/plugins/capacitor-theme-setter'
-import { getAddExpenseDialogOptions } from '../refuels/models'
+import { getAddButtonDialogOptions } from '../../layouts/models'
 
 const $q = useQuasar()
 $q.dark.set('auto')
@@ -98,7 +98,7 @@ const router = useRouter()
 const dashboardStore = useDashboardStore()
 const settingsStore = useSettingsStore()
 const mainLayoutStore = useMainLayoutStore()
-const { addExpenseDialogOptions } = getAddExpenseDialogOptions()
+const { addExpenseDialogOptions } = getAddButtonDialogOptions()
 const { t } = useI18n({ useScope: 'local', messages })
 
 const showChart = ref(false)
@@ -153,10 +153,6 @@ async function init() {
   await initSettings()
   mainLayoutStore.addButton.action = () =>
     optionsDialog(addExpenseDialogOptions)
-  router.beforeEach(() => {
-    mainLayoutStore.addButton.action = () =>
-      optionsDialog(addExpenseDialogOptions)
-  })
 }
 
 onMounted(async () => {
