+++ src/pages/refuels/ExpenseForm.vue
@@ -64,7 +64,7 @@
 </template>
 
 <script setup lang="ts">
-import { onMounted, onBeforeUnmount, toRaw, reactive, ref } from 'vue'
+import { onMounted, toRaw, reactive, ref } from 'vue'
 import { date as QuasarDate } from 'quasar'
 import { useRouter } from 'vue-router'
 import CInput from 'src/components/inputs/CInput.vue'
@@ -200,9 +200,5 @@ onMounted(async () => {
     expense.description = r.description.toString()
     expense.payedAmount = r.payedAmount.toString()
   }
-})
-
-onBeforeUnmount(() => {
-  mainLayoutStore.addButton.disabled = false
 })
 </script>
